Refactor to JSONtext VI package (BETA 1.5.0)
JSONtext is currently included as subtree, which means that users have
to pull changes from the main repository at bitbucket [1]. This will
inevitable cause issues with multiple different versions of JSONtext
and potential crosslinking between the subtree and official packages.

There is now an official BETA version with class serialization support
available on ni.com [2]. This commit refactors the project to use the
JSONtext VI package instead of the subtree (which is removed).

WARNING

BETA 1.5.0 has a known bug that will cause some unit tests to fail.
A workaround is provided on bitbucket [3].

Closes #3

[1] https://bitbucket.org/drjdpowell/jsontext
[2] https://forums.ni.com/t5/JDP-Science-Tools/BETA-version-of-JSONtext
-1-5-0/m-p/4136112
[3] https://bitbucket.org/drjdpowell/jsontext/issues/56/parse-json-from
-to-labview-objects#comment-60193488