Add package configuration
This project relies on multiple packages that are installed with the
VI Package Manager. To make this process easier and ensure that users
have the right versions installed, a VI Package Configuration file
should be included with the project.

The .vipc file requires VI Package Manager 2017 or newer.

Minimum required versions:
- JSONtext 1.5.0
- Caraya Unit Test Framework 1.1.0

Automatic dependencies:
- JDP Science Common Utilities 1.2.2
- JKI State Machine 2018.0.7
- jki_rsc_toolkits_palette 1.1

Closes #6