Revert large object handling from 5190b7a
Objects with more than 150 individual elements in their combined class
hierarchy will produce incorrect output with JSONtext.

Because of a limitation in '<vi.lib>\JDP Science\JDP Science Common
Utilities\Array of Variants to Cluster.vi' only the first 150 elements
will appear in the JSON string.

While it is certainly possible to reach this limit with deep class
hierarchies, it is easier to just update the function. This will also
make it consistent throughout JSONtext.

My implementation was trying to solve a problem for the sake of solving
the problem, not because of any real-world requirements (if my classes
ever reach this level of complexity I'm probably better off re-thinking
my strategy anyway).

As such I'm removing it for the sake of readability and consistency.