Fix pretty print does not work when type name handling is enabled
When type name handling is enabled, the pretty print status does not
affect the JSON output.

The problem is that type name handling currently ignores pretty print
and simply adds the value. Additionally, the Merge function clears
any formatting and returns a minified version of whatever JSON value
is provided as input.

This commit simply uses the Reformat function to forcibly reformat the
output. This will obviously become very slow for complex class
hierarchies, as the function is called for every object.

Fixes #11