Remove GCentral API example
I want to have much simpler examples to demonstrate the library. Even
though the GCentral API is a good example, it depends on an external
API that might change any minute. It is also not a good idea to ship
examples that utilize external APIs outside of their intended use-case.