Add examples
All examples are installed with the package and exposed by Example
Finder. Developers need to take care of the bin3 file, which must be
edited in the <labview>\examples\exbins folder and copied to the repo
manually (VIPM does not create or update it automatically).

The process is also explained at ni.com[1]

[1] https://forums.ni.com/t5/Developer-Center-Resources/Including-Examp
les-with-a-VI-Package-File/ta-p/3533408