Add JSON Serializable API
The JSON Serializable API defines a base class for serializable classes
('JSON Serializable'), children of which are serialized using methods
that they implement ('To JSON' and 'From JSON').

The serializer will only serialize classes that inherit from the JSON
Serializable class. Any other class will have value 'null' in the JSON
output and always returns the base type.

The type of the object is always added with key "$type" in the JSON
output (for any class). If a child class returns any non-object JSON
value, it is stored with key "$data".

Example

  {
    "$type":"My JSON Serializable.lvclass",
    "$data":[1,2,3]
  }