Use Variant to Flattened String (special) from JSONtext
JSONtext 1.6.5 provides a new function that solves an issue, where an
array (or cluster) of objects would fail to unflatten from string with
error 1400.

This error occurs when attempting to rebuild an object from variant
data, when the variant data contains other variant elements. In this
case the type information of the variant data does not automatically
translate into the desired object data. This is only an issue for
compound and array types.

For example, an array of object `array[<object>]` would appear as
`array[<variant<object>>]`, which has entirely different type data.

'Variant to Flattened String (special)' is a function implemented by
JSONtext to address this issue by unpacking variant type information
and returning only the raw type data, which effectively solves this
issue.

The function is private to JSONtext, which is why this commit imports
the function into the vendor directory of this project with the license
text added to the VI description.

Closes #20