Fix inline type name handling without data member
This addresses an error that occurs when attempting to deserialize
classes without data member (empty private data cluster). In these
cases, JSONtext is unable to deserialize the private data cluster
because it contains no elements (which is not possible for any other
kind of cluster).

Fixed by checking the number of data elements in the private data
cluster before parsing. For empty data clusters parsing is skipped and
the data cluster is used as is.

References #21