Fix numeric types with less than 32 bits are wrong
This adds additional unit tests for numeric types when using the
Generic Object Serializer.

Numeric types with less than 32 bits currently fail. This includes:
- Enum U8
- Enum U16
- U8
- U16
- I8
- I16

These tests fail because the serializer assumes that type information
is maintained by JSONtext when parsing from JSON. This, however, is
not the case with certain numeric types, for which the parser always
returns 32-bit values, no matter which input type is specified.

When the 32-bit value is later combined with the type information of
a type with less bits, the lower bits are truncated, which results in
incorrect values.

Note that 64-bit values are not affected because they are handled in
native bitness by JSONtext.

Fixed by changing the minimum required version for JSONtext to 1.6.7,
which returns the correct bitness for all numeric types.

References #22